Removed custom Admin query lifecycle
no ref

The canonical query migration only needs shared query options, ordinary TanStack observers, and a reset boundary. Removing custom cache scopes and request-policy tracking keeps recovery aligned with standard TanStack Query behavior.
