+++ apps/admin-x-framework/src/api/current-user.ts
@@ -1,7 +1,5 @@
 import { queryOptions, useQuery, useSuspenseQuery } from '@tanstack/react-query';
 import { apiUrl, useFetchApi } from '../utils/api/fetch-api';
-import { SETTINGS_BOOTSTRAP_QUERY_SCOPE } from '../utils/api/query-scopes';
-import { withQueryErrorPolicy } from '../utils/api/query-error-policy';
 import { UsersResponseType } from './users';
 
 export const usersDataType = 'UsersResponseType';
@@ -14,12 +12,7 @@ export const useCurrentUserQueryOptions = () => {
 
   return queryOptions({
     queryKey: currentUserQueryKey,
-    queryFn: (context) =>
-      withQueryErrorPolicy(context, true, () => fetchApi<UsersResponseType>(currentUserUrl)),
-    meta: {
-      defaultErrorHandler: true,
-      errorResetScope: SETTINGS_BOOTSTRAP_QUERY_SCOPE,
-    },
+    queryFn: () => fetchApi<UsersResponseType>(currentUserUrl),
   });
 };
 
@@ -45,10 +38,6 @@ export const useCurrentUserSuspense = () => {
     ...currentUserQueryOptions,
     select: (data) => data.users[0],
   });
-
-  if (result.error && !result.isFetching) {
-    throw result.error;
-  }
 
   return result;
 };
