+++ apps/admin-x-framework/src/api/settings.ts
@@ -1,5 +1,4 @@
 import { Meta, createMutation, createQueryResource } from '../utils/api/hooks';
-import { SETTINGS_BOOTSTRAP_QUERY_SCOPE } from '../utils/api/query-scopes';
 import { Config } from './config';
 
 // Types
@@ -33,7 +32,6 @@ const dataType = 'SettingsResponseType';
 const browseSettingsQuery = {
   dataType,
   path: '/settings/',
-  errorResetScope: SETTINGS_BOOTSTRAP_QUERY_SCOPE,
   defaultSearchParams: {
     group:
       'site,theme,private,members,portal,newsletter,email,labs,slack,unsplash,views,firstpromoter,editor,comments,analytics,announcement,pintura,donations,security,social_web,explore,transistor',
