+++ apps/admin-x-framework/src/api/site.ts
@@ -1,5 +1,4 @@
 import { createQueryResource } from '../utils/api/hooks';
-import { SETTINGS_BOOTSTRAP_QUERY_SCOPE } from '../utils/api/query-scopes';
 import { Config, hasSendingDomain, isManagedEmail, sendingDomain } from './config';
 
 // Types
@@ -28,7 +27,6 @@ const dataType = 'SiteResponseType';
 const browseSiteResource = createQueryResource<SiteResponseType>({
   dataType,
   path: '/site/',
-  errorResetScope: SETTINGS_BOOTSTRAP_QUERY_SCOPE,
 });
 
 export const useBrowseSite = browseSiteResource.useQuery;
