+++ apps/admin-x-framework/src/providers/framework-provider.tsx
@@ -1,24 +1,7 @@
 import { ErrorBoundary as SentryErrorBoundary } from '@sentry/react';
-import {
-  Query,
-  QueryCache,
-  QueryClient,
-  QueryClientProvider,
-  QueryErrorResetBoundary,
-} from '@tanstack/react-query';
-import { ReactNode, createContext, useContext, useLayoutEffect, useMemo, useRef } from 'react';
-import { handleFrameworkError } from '../utils/handle-error';
+import { QueryClient, QueryClientProvider, QueryErrorResetBoundary } from '@tanstack/react-query';
+import { ReactNode, createContext, useContext, useMemo } from 'react';
 import queryClient from '../utils/query-client';
-import { clearQueryErrorPolicy, takeQueryErrorPolicy } from '../utils/api/query-error-policy';
-
-declare module '@tanstack/react-query' {
-  interface Register {
-    queryMeta: {
-      defaultErrorHandler?: boolean;
-      errorResetScope?: string;
-    };
-  }
-}
 
 export type ExternalLink = {
   isExternal: true;
@@ -98,62 +81,6 @@ const FrameworkContext = createContext<FrameworkContextType>({
   onDelete: () => {},
 });
 
-type QueryErrorHandler = (
-  error: unknown,
-  query: Query<unknown, unknown, unknown, readonly unknown[]>,
-) => void;
-type QueryCacheRegistration = {
-  handlers: Map<symbol, QueryErrorHandler>;
-  originalHandler: QueryCache['config']['onError'];
-  unsubscribe: () => void;
-};
-
-const queryCacheRegistrations = new WeakMap<QueryCache, QueryCacheRegistration>();
-
-function registerQueryErrorHandler(cache: QueryCache, token: symbol, handler: QueryErrorHandler) {
-  let registration = queryCacheRegistrations.get(cache);
-
-  if (!registration) {
-    registration = {
-      handlers: new Map(),
-      originalHandler: cache.config.onError,
-      unsubscribe: cache.subscribe((event) => {
-        if (event.type === 'updated' && event.action.type === 'fetch') {
-          // Cancellation can settle TanStack's retryer before the underlying
-          // request promise. Clear its snapshot before any subsequent queryFn
-          // records the policy for a new request on the same Query.
-          clearQueryErrorPolicy(cache, event.query);
-        }
-      }),
-    };
-    queryCacheRegistrations.set(cache, registration);
-
-    cache.config.onError = (error, query) => {
-      registration?.originalHandler?.(error, query);
-      // Several embedded apps can temporarily mount FrameworkProvider against
-      // the shared client. Reporting is a cache-level policy, so one provider
-      // handles each failure rather than duplicating the same toast/Sentry event.
-      registration?.handlers.values().next().value?.(error, query);
-    };
-  }
-
-  registration.handlers.set(token, handler);
-}
-
-function unregisterQueryErrorHandler(cache: QueryCache, token: symbol) {
-  const registration = queryCacheRegistrations.get(cache);
-  if (!registration) {
-    return;
-  }
-
-  registration.handlers.delete(token);
-  if (registration.handlers.size === 0) {
-    registration.unsubscribe();
-    cache.config.onError = registration.originalHandler;
-    queryCacheRegistrations.delete(cache);
-  }
-}
-
 export function FrameworkProvider({
   children,
   queryClient: queryClientOverride,
@@ -166,22 +93,6 @@ export function FrameworkProvider({
 
     return queryClient;
   }, [queryClientOverride]);
-
-  const queryErrorHandlerToken = useRef(Symbol('framework-query-error-handler')).current;
-  const cache = client.getQueryCache();
-  useLayoutEffect(() => {
-    registerQueryErrorHandler(cache, queryErrorHandlerToken, (error, query) => {
-      // Canonical resource queryFns snapshot the initiating request's policy;
-      // the public Query.meta fallback covers imperative/non-resource queries.
-      const shouldHandle =
-        takeQueryErrorPolicy(cache, query) ?? query.meta?.defaultErrorHandler === true;
-      if (shouldHandle) {
-        handleFrameworkError(error, { sentryDSN: props.sentryDSN });
-      }
-    });
-
-    return () => unregisterQueryErrorHandler(cache, queryErrorHandlerToken);
-  }, [cache, props.sentryDSN, queryErrorHandlerToken]);
 
   return (
     <SentryErrorBoundary>
