--- apps/admin-x-framework/src/utils/api/query-scopes.ts
@@ -1,1 +1,0 @@
-export const SETTINGS_BOOTSTRAP_QUERY_SCOPE = 'settings-bootstrap';
